+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/InvoiceTablePanel.java
@@ -21,9 +21,9 @@ package is.codion.framework.demos.chinook.ui;
 import is.codion.common.model.condition.ColumnConditions;
 import is.codion.framework.demos.chinook.domain.Chinook.Invoice;
 import is.codion.framework.domain.entity.attribute.Attribute;
-import is.codion.swing.common.ui.component.table.ColumnConditionsPanel;
 import is.codion.swing.common.ui.component.table.ConditionPanel;
 import is.codion.swing.common.ui.component.table.FilterTableColumnModel;
+import is.codion.swing.common.ui.component.table.TableConditionPanel;
 import is.codion.swing.framework.model.SwingEntityTableModel;
 import is.codion.swing.framework.ui.EntityTablePanel;
 
@@ -37,11 +37,11 @@ public final class InvoiceTablePanel extends EntityTablePanel {
 	public InvoiceTablePanel(SwingEntityTableModel tableModel) {
 		super(tableModel, config -> config
 						.editable(attributes -> attributes.remove(Invoice.TOTAL))
-						.columnConditionsPanelFactory(new InvoiceConditionPanelFactory(tableModel))
+						.conditionPanelFactory(new InvoiceConditionPanelFactory(tableModel))
 						.conditionState(SIMPLE));
 	}
 
-	private static final class InvoiceConditionPanelFactory implements ColumnConditionsPanel.Factory<Attribute<?>> {
+	private static final class InvoiceConditionPanelFactory implements TableConditionPanel.Factory<Attribute<?>> {
 
 		private final SwingEntityTableModel tableModel;
 
@@ -50,10 +50,10 @@ public final class InvoiceTablePanel extends EntityTablePanel {
 		}
 
 		@Override
-		public ColumnConditionsPanel<Attribute<?>> create(ColumnConditions<Attribute<?>> conditionModel,
-																											Map<Attribute<?>, ConditionPanel<?>> conditionPanels,
-																											FilterTableColumnModel<Attribute<?>> columnModel,
-																											Consumer<ColumnConditionsPanel<Attribute<?>>> onPanelInitialized) {
+		public TableConditionPanel<Attribute<?>> create(ColumnConditions<Attribute<?>> conditionModel,
+																										Map<Attribute<?>, ConditionPanel<?>> conditionPanels,
+																										FilterTableColumnModel<Attribute<?>> columnModel,
+																										Consumer<TableConditionPanel<Attribute<?>>> onPanelInitialized) {
 			return new InvoiceConditionPanel(tableModel, conditionPanels, columnModel, onPanelInitialized);
 		}
 	}