+++ changelog.md
@@ -23,6 +23,7 @@ Codion Change Log
 - ColumnConditionsPanel captions function added as constructor parameter.
 - ColumnConditionPanel renamed ConditionPanel.
 - ColumnConditionsPanel.panels() and panel() renamed get().
+- ColumnConditionsPanel renamed TableConditionPanel.
 ### is.codion.framework.domain
 - Entity.copy(), copyBuilder() and deepCopy() moved to Entity.Copy, renamed mutable(), builder() and deep() respectively, immutable() added.
 - ImmutableEntity bug fixed, StackOverflowError when creating an immutable entity with cyclical foreign key dependencies, some refactoring.